requirements.txt
langchain
langchain-community
langchain-ollama
langchain-groq
langchain-core
langchain-text-splitters
chromadb
beautifulsoup4
requests
streamlit
pypdf
python-dotenv
groq
huggingface-hub